Add Pi-hole (nameserver) host and split up inventory
Since we got a new rpi, we can split up some tasks and distribute the
roles over the different hosts.
For now, this is running Pi-hole on one of the rpi's, but this seems
like a good first step to have a set up whereby we can start managing
multiple hosts via one repo.