Remove obsolete script to write SD
Just a quick search, but seems that the default admin user/password
login is no longer available; one has to create the user upon first
boot.  This is cumbersome, but fortunately, it is still possible to
create a default administrator with the Raspberry Pi Imager application.
Possibly there's still an automated way (I'd be surprised if it isn't)
but it's time to just get this thing basically up&running and accessible
via SSH on the internal network.