Better name for the mount holding data
'cache' was bad... it was suggested because of Jellyfin's cache dir
living on that drive, but there's other stuff too, and all things
Jellyfin go in a subdir, so renaming it, not that 'data' is so much
better, but it _is_ better.
Leaving the Jellyfin provisioning out of this commit, because I'm still
working on that.